+++ chalice/config.py
@@ -230,6 +230,18 @@ class Config(object):
                                   varies_per_chalice_stage=True)
 
     @property
+    def api_gateway_endpoint_vpce(self):
+        # type: () -> str
+        return self._chain_lookup('api_gateway_endpoint_vpce',
+                                  varies_per_chalice_stage=True)
+
+    @property
+    def api_gateway_policy(self):
+        # type: () -> Optional[Dict]
+        return self._chain_lookup('api_gateway_policy',
+                                  varies_per_chalice_stage=True)
+
+    @property
     def minimum_compression_size(self):
         # type: () -> int
         return self._chain_lookup('minimum_compression_size',