+++ chalice/deploy/validate.py
@@ -46,13 +46,26 @@ def validate_configuration(config):
     validate_unique_function_names(config)
     validate_feature_flags(config.chalice_app)
     validate_endpoint_type(config)
+    validate_resource_policy(config)
+
+
+def validate_resource_policy(config):
+    # type: (Config) -> None
+    if config.api_gateway_endpoint_type != 'PRIVATE':
+        return
+    if config.api_gateway_policy:
+        return
+    if not config.api_gateway_endpoint_vpce:
+        raise ValueError(
+            ("Private Endpoints require api_gateway_policy or "
+             "api_gateway_endpoint_vpce specified"))
 
 
 def validate_endpoint_type(config):
     # type: (Config) -> None
     if not config.api_gateway_endpoint_type:
         return
-    valid_types = ('EDGE', 'REGIONAL')
+    valid_types = ('EDGE', 'REGIONAL', 'PRIVATE')
     if config.api_gateway_endpoint_type not in valid_types:
         raise ValueError(
             "api gateway endpoint type must be one of %s" % (