+++ tests/unit/deploy/test_validate.py
@@ -13,6 +13,7 @@ from chalice.deploy.validate import validate_route_content_types
 from chalice.deploy.validate import validate_unique_function_names
 from chalice.deploy.validate import validate_feature_flags
 from chalice.deploy.validate import validate_endpoint_type
+from chalice.deploy.validate import validate_resource_policy
 from chalice.deploy.validate import ExperimentalFeatureError
 
 
@@ -234,6 +235,25 @@ def test_can_validate_updated_custom_binary_types(sample_app):
                                         sample_app.api.binary_types) is None
 
 
+def test_can_validate_resource_policy(sample_app):
+    config = Config.create(
+        chalice_app=sample_app, api_gateway_endpoint_type='PRIVATE')
+    with pytest.raises(ValueError):
+        validate_resource_policy(config)
+
+    config = Config.create(
+        chalice_app=sample_app,
+        api_gateway_endpoint_vpce='vpce-abc123',
+        api_gateway_endpoint_type='PRIVATE')
+    validate_resource_policy(config)
+
+    config = Config.create(
+        chalice_app=sample_app,
+        api_gateway_policy={'Statement': []},
+        api_gateway_endpoint_type='PRIVATE')
+    validate_resource_policy(config)
+
+
 def test_can_validate_endpoint_type(sample_app):
     config = Config.create(
         chalice_app=sample_app, api_gateway_endpoint_type='EDGE2')
@@ -243,11 +263,6 @@ def test_can_validate_endpoint_type(sample_app):
     config = Config.create(
         chalice_app=sample_app, api_gateway_endpoint_type='REGIONAL')
     validate_endpoint_type(config)
-
-    config = Config.create(
-        chalice_app=sample_app, api_gateway_endpoint_type='PRIVATE')
-    with pytest.raises(ValueError):
-        validate_endpoint_type(config)
 
 
 def test_can_validate_feature_flags(sample_app):